+++ packages/typespec-mcp-http-server-js/package.json
@@ -36,7 +36,9 @@
   "type": "module",
   "devDependencies": {
     "@alloy-js/cli": "catalog:",
+    "@alloy-js/rollup-plugin": "catalog:",
     "@types/node": "catalog:",
-    "concurrently": "catalog:"
+    "concurrently": "catalog:",
+    "vitest": "catalog:"
   }
 }
