+++ packages/typespec-mcp-server-js/package.json
@@ -40,6 +40,7 @@
     "@alloy-js/cli": "catalog:",
     "@alloy-js/rollup-plugin": "catalog:",
     "@types/node": "catalog:",
+    "vitest": "catalog:",
     "concurrently": "catalog:"
   }
 }
