+++ packages/typespec-mcp-server-js/test/test-host.ts
@@ -1,0 +1,12 @@
+import { createTestHost, createTestWrapper } from "@typespec/compiler/testing";
+import { TypeSpecMcp } from "typespec-mcp/testing";
+
+export async function createMcpTestHost() {
+  return createTestHost({
+    libraries: [TypeSpecMcp],
+  });
+}
+export async function createTestRunner() {
+  const host = await createMcpTestHost();
+  return createTestWrapper(host, { autoUsings: ["MCP"] });
+}
